+++ src/parse/check_type.rs
@@ -3,14 +3,14 @@ pub fn string(value: &str) -> bool {
 }
 
 pub fn int(value: &str) -> bool {
-    value.parse::<i64>().is_ok()
+    value.parse::<i128>().is_ok()
 }
 
 pub fn unsigned(mut value: &str) -> bool {
     if value.starts_with("(unsigned)") {
         value = value[10..].trim();
     }
-    value.parse::<u64>().is_ok()
+    value.parse::<u128>().is_ok()
 }
 
 pub fn float(value: &str) -> bool {
